my-chatbot/requirements.txt
flask
langchain
pypdf
sentence-transformers
faiss-cpu
flask-cors